requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
httpx
python-multipart